Added files and modified ProgressContainer.js and PC.js
