+++ packages/progress-container/package.json
@@ -4,14 +4,6 @@
   "description": "Prifina Progress Container widget",
   "author": "Jenn K <jmjenniferkim@gmail.com>",
   "license": "MIT",
-  "main": "dist/esm/index.js",
-  "publishConfig": {
-    "access": "public"
-  },
-  "files": [
-    "dist"
-  ],
-  "dependencies": {},
   "devDependencies": {
     "@babel/cli": "^7.12.13",
     "@babel/core": "^7.12.13",
@@ -20,7 +12,9 @@
     "@babel/preset-env": "^7.12.13",
     "@babel/preset-react": "^7.12.13",
     "@babel/runtime": "^7.12.13",
-    "@prifina/hooks": "^0.0.48",
+    "@blend-ui/core": "^1.1.57-alpha.0",
+    "@prifina/faker": "latest",
+    "@prifina/hooks": "latest",
     "@storybook/addon-actions": "^6.1.17",
     "@storybook/addon-docs": "^6.1.17",
     "@storybook/addon-knobs": "^6.1.17",
@@ -47,9 +41,7 @@
     "webpack-assets-manifest": "^5.0.1",
     "webpack-bundle-analyzer": "^4.4.0",
     "webpack-cli": "^4.5.0",
-    "webpack-dev-server": "^3.11.2",
-    "https-browserify": "^1.0.0",
-    "stream-http": "^3.1.1"
+    "webpack-dev-server": "^3.11.2"
   },
   "scripts": {
     "build": "cross-env NODE_ENV=production webpack --mode production --config config/webpack.config.js",
@@ -58,9 +50,28 @@
     "storybook": "start-storybook -p 9009 -s public"
   },
   "peerDependencies": {
-    "@prifina/hooks": "^0.0.48",
+    "@prifina/faker": "latest",
+    "@prifina/hooks": "latest",
     "react": "^17.0.1",
     "styled-components": "^5.2.1",
     "styled-system": "^5.1.5"
+  },
+  "dependencies": {
+    "@blend-ui/core": "^1.1.57-alpha.0",
+    "@blend-ui/icons": "^1.1.8-alpha.0",
+    "@blend-ui/progress": "^1.0.7-alpha.0",
+    "@chakra-ui/react": "^1.3.3",
+    "@chakra-ui/theme": "^1.7.0",
+    "@chakra-ui/theme-tools": "^1.1.0",
+    "@emotion/react": "^11.1.5",
+    "@emotion/styled": "^11.1.5",
+    "@styled-system/prop-types": "^5.1.5",
+    "@styled-system/theme-get": "^5.1.2",
+    "framer-motion": "^3.10.0",
+    "react-router-dom": "^5.2.0",
+    "react-select": "^4.2.1",
+    "recharts": "^2.0.8"
   }
 }
+
+
