+++ packages/progress-container/src/Error.js
@@ -1,0 +1,16 @@
+import React from "react";
+
+import { CircularProgress, CircularProgressLabel } from "@blend-ui/progress";
+import { ThemeProvider, Text, Avatar, Flex } from "@blend-ui/core";
+import { BlendIcon } from "@blend-ui/icons";
+import Box from "@blend-ui/core/dist/esm/Box";
+
+function Error() {
+  return (
+    <Box>
+      <Text>Error</Text>
+    </Box>
+  );
+}
+
+export default Error;
