+++ packages/progress-container/src/data.js
@@ -1,0 +1,74 @@
+export const data = [
+  {
+    name: "Day 5",
+    uv: 4000,
+    Calories: 2400,
+    amt: 2400,
+  },
+  {
+    name: "Day 10",
+    uv: 3000,
+    Calories: 1398,
+    amt: 2210,
+  },
+  {
+    name: "Day 15",
+    uv: 2000,
+    Calories: 3800,
+    amt: 2290,
+  },
+  {
+    name: "Day 20",
+    uv: 2780,
+    Calories: 3908,
+    amt: 2000,
+  },
+  {
+    name: "Day 25",
+    uv: 1890,
+    Calories: 4800,
+    amt: 2181,
+  },
+  {
+    name: "Day 30",
+    uv: 2390,
+    Calories: 3800,
+    amt: 2500,
+  },
+  {
+    name: "Day 35",
+    uv: 3490,
+    Calories: 4300,
+    amt: 2100,
+  },
+];
+
+const fitbit = [
+  {
+    name: "Day 5",
+    uv: 4000,
+    Calories: 2400,
+    amt: 2400,
+  },
+  {
+    name: "Day 10",
+    uv: 3000,
+    Calories: 1398,
+    amt: 2210,
+  },
+];
+
+const oura = [
+  {
+    name: "Day 5",
+    uv: 4000,
+    Calories: 2400,
+    amt: 2400,
+  },
+  {
+    name: "Day 10",
+    uv: 3000,
+    Calories: 1398,
+    amt: 2210,
+  },
+];
