+++ packages/progress-container/src/data/data.js
@@ -1,0 +1,1110 @@
+export function Device(device) {
+  let result;
+
+  switch (device) {
+    case "fitbit":
+      result = [
+        {
+          name: "Feb-6",
+          totalCalories: 1642,
+          activeCalories: 0,
+          totalSteps: 0,
+          totalDistance: 0,
+          inactiveMinutes: 1440,
+          lowActiveMinutes: 0,
+          mediumActiveMinutes: 0,
+          highActiveMinutes: 0,
+          lightSleepTime: 232,
+          deepSleepTime: 83,
+          timeSpentInBed: 449,
+          totalSleepTime: 416,
+          REMSleepTime: 101,
+          totalAwakeTime: 33,
+        },
+        {
+          name: "Feb-7",
+          totalCalories: 1945,
+          activeCalories: 331,
+          totalSteps: 3270,
+          totalDistance: 2380,
+          inactiveMinutes: 1378,
+          lowActiveMinutes: 37,
+          mediumActiveMinutes: 4,
+          highActiveMinutes: 21,
+          lightSleepTime: 205,
+          deepSleepTime: 63,
+          timeSpentInBed: 402,
+          totalSleepTime: 362,
+          REMSleepTime: 94,
+          totalAwakeTime: 40,
+        },
+        {
+          name: "Feb-8",
+          totalCalories: 2859,
+          activeCalories: 1345,
+          totalSteps: 15000,
+          totalDistance: 13.55 * 1000,
+          inactiveMinutes: 804,
+          lowActiveMinutes: 98,
+          mediumActiveMinutes: 3,
+          highActiveMinutes: 86,
+          lightSleepTime: 218,
+          deepSleepTime: 61,
+          timeSpentInBed: 421,
+          totalSleepTime: 382,
+          REMSleepTime: 103,
+          totalAwakeTime: 39,
+        },
+        {
+          name: "Feb-9",
+          totalCalories: 2932,
+          activeCalories: 1431,
+          totalSteps: 15756,
+          totalDistance: 14.15 * 1000,
+          inactiveMinutes: 844,
+          lowActiveMinutes: 95,
+          mediumActiveMinutes: 7,
+          highActiveMinutes: 92,
+          lightSleepTime: 165,
+          deepSleepTime: 94,
+          timeSpentInBed: 437,
+          totalSleepTime: 387,
+          REMSleepTime: 128,
+          totalAwakeTime: 50,
+        },
+        {
+          name: "Feb-10",
+          totalCalories: 3466,
+          activeCalories: 2070,
+          totalSteps: 25218,
+          totalDistance: 21.02 * 1000,
+          inactiveMinutes: 742,
+          lowActiveMinutes: 97,
+          mediumActiveMinutes: 9,
+          highActiveMinutes: 171,
+          lightSleepTime: 203,
+          deepSleepTime: 74,
+          timeSpentInBed: 418,
+          totalSleepTime: 371,
+          REMSleepTime: 94,
+          totalAwakeTime: 47,
+        },
+        {
+          name: "Feb-11",
+          totalCalories: 2954,
+          activeCalories: 1476,
+          totalSteps: 16247,
+          totalDistance: 14.45 * 1000,
+          inactiveMinutes: 786,
+          lowActiveMinutes: 127,
+          mediumActiveMinutes: 7,
+          highActiveMinutes: 83,
+          lightSleepTime: 171,
+          deepSleepTime: 99,
+          timeSpentInBed: 415,
+          totalSleepTime: 376,
+          REMSleepTime: 106,
+          totalAwakeTime: 39,
+        },
+        {
+          name: "Feb-12",
+          totalCalories: 2903,
+          activeCalories: 1393,
+          totalSteps: 15520,
+          totalDistance: 13.96 * 1000,
+          inactiveMinutes: 830,
+          lowActiveMinutes: 98,
+          mediumActiveMinutes: 7,
+          highActiveMinutes: 87,
+          lightSleepTime: 181,
+          deepSleepTime: 100,
+          timeSpentInBed: 435,
+          totalSleepTime: 397,
+          REMSleepTime: 116,
+          totalAwakeTime: 38,
+        },
+        {
+          name: "Feb-13",
+          totalCalories: 3900,
+          activeCalories: 2612,
+          totalSteps: 27972,
+          totalDistance: 20.32 * 1000,
+          inactiveMinutes: 641,
+          lowActiveMinutes: 118,
+          mediumActiveMinutes: 51,
+          highActiveMinutes: 215,
+          lightSleepTime: 216,
+          deepSleepTime: 53,
+          timeSpentInBed: 414,
+          totalSleepTime: 380,
+          REMSleepTime: 111,
+          totalAwakeTime: 34,
+        },
+        {
+          name: "Feb-14",
+          totalCalories: 3009,
+          activeCalories: 1468,
+          totalSteps: 15577,
+          totalDistance: 13.95 * 1000,
+          inactiveMinutes: 788,
+          lowActiveMinutes: 110,
+          mediumActiveMinutes: 16,
+          highActiveMinutes: 91,
+          lightSleepTime: 189,
+          deepSleepTime: 64,
+          timeSpentInBed: 406,
+          totalSleepTime: 359,
+          REMSleepTime: 106,
+          totalAwakeTime: 47,
+        },
+        {
+          name: "Feb-15",
+          totalCalories: 2981,
+          activeCalories: 1477,
+          totalSteps: 15492,
+          totalDistance: 13.86 * 1000,
+          inactiveMinutes: 816,
+          lowActiveMinutes: 119,
+          mediumActiveMinutes: 6,
+          highActiveMinutes: 85,
+          lightSleepTime: 215,
+          deepSleepTime: 64,
+          timeSpentInBed: 431,
+          totalSleepTime: 369,
+          REMSleepTime: 90,
+          totalAwakeTime: 62,
+        },
+        {
+          name: "Feb-16",
+          totalCalories: 2892,
+          activeCalories: 1404,
+          totalSteps: 15645,
+          totalDistance: 14.05 * 1000,
+          inactiveMinutes: 829,
+          lowActiveMinutes: 115,
+          mediumActiveMinutes: 4,
+          highActiveMinutes: 86,
+          lightSleepTime: 197,
+          deepSleepTime: 70,
+          timeSpentInBed: 406,
+          totalSleepTime: 369,
+          REMSleepTime: 102,
+          totalAwakeTime: 37,
+        },
+        {
+          name: "Feb-17",
+          totalCalories: 2544,
+          activeCalories: 999,
+          totalSteps: 11568,
+          totalDistance: 11.08 * 1000,
+          inactiveMinutes: 894,
+          lowActiveMinutes: 47,
+          mediumActiveMinutes: 2,
+          highActiveMinutes: 66,
+          lightSleepTime: 213,
+          deepSleepTime: 38,
+          timeSpentInBed: 348,
+          totalSleepTime: 314,
+          REMSleepTime: 63,
+          totalAwakeTime: 34,
+        },
+        {
+          name: "Feb-18",
+          totalCalories: 3780,
+          activeCalories: 2434,
+          totalSteps: 28456,
+          totalDistance: 23.29 * 1000,
+          inactiveMinutes: 720,
+          lowActiveMinutes: 92,
+          mediumActiveMinutes: 16,
+          highActiveMinutes: 206,
+          lightSleepTime: 264,
+          deepSleepTime: 81,
+          timeSpentInBed: 485,
+          totalSleepTime: 430,
+          REMSleepTime: 85,
+          totalAwakeTime: 55,
+        },
+        {
+          name: "Feb-19",
+          totalCalories: 2873,
+          activeCalories: 1362,
+          totalSteps: 15664,
+          totalDistance: 14.1 * 1000,
+          inactiveMinutes: 909,
+          lowActiveMinutes: 86,
+          mediumActiveMinutes: 5,
+          highActiveMinutes: 92,
+          lightSleepTime: 259,
+          deepSleepTime: 84,
+          timeSpentInBed: 500,
+          totalSleepTime: 442,
+          REMSleepTime: 99,
+          totalAwakeTime: 58,
+        },
+        {
+          name: "Feb-20",
+          totalCalories: 3787,
+          activeCalories: 2418,
+          totalSteps: 27602,
+          totalDistance: 20.05 * 1000,
+          inactiveMinutes: 594,
+          lowActiveMinutes: 120,
+          mediumActiveMinutes: 21,
+          highActiveMinutes: 220,
+          lightSleepTime: 187,
+          deepSleepTime: 62,
+          timeSpentInBed: 394,
+          totalSleepTime: 356,
+          REMSleepTime: 107,
+          totalAwakeTime: 38,
+        },
+        {
+          name: "Feb-21",
+          totalCalories: 3039,
+          activeCalories: 1505,
+          totalSteps: 15523,
+          totalDistance: 13.97 * 1000,
+          inactiveMinutes: 729,
+          lowActiveMinutes: 108,
+          mediumActiveMinutes: 10,
+          highActiveMinutes: 93,
+          lightSleepTime: 166,
+          deepSleepTime: 115,
+          timeSpentInBed: 427,
+          totalSleepTime: 368,
+          REMSleepTime: 87,
+          totalAwakeTime: 59,
+        },
+        {
+          name: "Feb-22",
+          totalCalories: 2873,
+          activeCalories: 1381,
+          totalSteps: 15196,
+          totalDistance: 13.64 * 1000,
+          inactiveMinutes: 853,
+          lowActiveMinutes: 97,
+          mediumActiveMinutes: 5,
+          highActiveMinutes: 91,
+          lightSleepTime: 238,
+          deepSleepTime: 47,
+          timeSpentInBed: 423,
+          totalSleepTime: 390,
+          REMSleepTime: 105,
+          totalAwakeTime: 33,
+        },
+        {
+          name: "Feb-23",
+          totalCalories: 3024,
+          activeCalories: 1522,
+          totalSteps: 15794,
+          totalDistance: 14.09 * 1000,
+          inactiveMinutes: 803,
+          lowActiveMinutes: 100,
+          mediumActiveMinutes: 16,
+          highActiveMinutes: 94,
+          lightSleepTime: 223,
+          deepSleepTime: 93,
+          timeSpentInBed: 462,
+          totalSleepTime: 408,
+          REMSleepTime: 92,
+          totalAwakeTime: 54,
+        },
+        {
+          name: "Feb-24",
+          totalCalories: 3591,
+          activeCalories: 2233,
+          totalSteps: 25937,
+          totalDistance: 21.52 * 1000,
+          inactiveMinutes: 701,
+          lowActiveMinutes: 131,
+          mediumActiveMinutes: 4,
+          highActiveMinutes: 181,
+          lightSleepTime: 236,
+          deepSleepTime: 58,
+          timeSpentInBed: 435,
+          totalSleepTime: 392,
+          REMSleepTime: 98,
+          totalAwakeTime: 43,
+        },
+        {
+          name: "Feb-25",
+          totalCalories: 2918,
+          activeCalories: 1378,
+          totalSteps: 14942,
+          totalDistance: 13.49 * 1000,
+          inactiveMinutes: 796,
+          lowActiveMinutes: 84,
+          mediumActiveMinutes: 6,
+          highActiveMinutes: 92,
+          lightSleepTime: 0,
+          deepSleepTime: 0,
+          timeSpentInBed: 91,
+          totalSleepTime: 89,
+          REMSleepTime: 0,
+          totalAwakeTime: 2,
+        },
+        {
+          name: "Feb-26",
+          totalCalories: 2849,
+          activeCalories: 1330,
+          totalSteps: 15147,
+          totalDistance: 13.65 * 1000,
+          inactiveMinutes: 737,
+          lowActiveMinutes: 80,
+          mediumActiveMinutes: 7,
+          highActiveMinutes: 90,
+          lightSleepTime: 187,
+          deepSleepTime: 52,
+          timeSpentInBed: 379,
+          totalSleepTime: 354,
+          REMSleepTime: 115,
+          totalAwakeTime: 25,
+        },
+        {
+          name: "Feb-27",
+          totalCalories: 3657,
+          activeCalories: 2326,
+          totalSteps: 27664,
+          totalDistance: 20.1 * 1000,
+          inactiveMinutes: 707,
+          lowActiveMinutes: 123,
+          mediumActiveMinutes: 14,
+          highActiveMinutes: 217,
+          lightSleepTime: 293,
+          deepSleepTime: 75,
+          timeSpentInBed: 482,
+          totalSleepTime: 435,
+          REMSleepTime: 67,
+          totalAwakeTime: 47,
+        },
+        {
+          name: "Feb-28",
+          totalCalories: 2912,
+          activeCalories: 1374,
+          totalSteps: 15128,
+          totalDistance: 13.62 * 1000,
+          inactiveMinutes: 764,
+          lowActiveMinutes: 97,
+          mediumActiveMinutes: 5,
+          highActiveMinutes: 92,
+          lightSleepTime: 187,
+          deepSleepTime: 54,
+          timeSpentInBed: 397,
+          totalSleepTime: 346,
+          REMSleepTime: 105,
+          totalAwakeTime: 51,
+        },
+        {
+          name: "Mar-1",
+          totalCalories: 2893,
+          activeCalories: 1408,
+          totalSteps: 15986,
+          totalDistance: 14.36 * 1000,
+          inactiveMinutes: 844,
+          lowActiveMinutes: 102,
+          mediumActiveMinutes: 9,
+          highActiveMinutes: 88,
+          lightSleepTime: 245,
+          deepSleepTime: 53,
+          timeSpentInBed: 459,
+          totalSleepTime: 401,
+          REMSleepTime: 103,
+          totalAwakeTime: 58,
+        },
+        {
+          name: "Mar-2",
+          totalCalories: 2810,
+          activeCalories: 1312,
+          totalSteps: 15097,
+          totalDistance: 13.64 * 1000,
+          inactiveMinutes: 801,
+          lowActiveMinutes: 85,
+          mediumActiveMinutes: 7,
+          highActiveMinutes: 88,
+          lightSleepTime: 266,
+          deepSleepTime: 44,
+          timeSpentInBed: 458,
+          totalSleepTime: 405,
+          REMSleepTime: 95,
+          totalAwakeTime: 53,
+        },
+        {
+          name: "Mar-3",
+          totalCalories: 3741,
+          activeCalories: 2429,
+          totalSteps: 28905,
+          totalDistance: 23.68 * 1000,
+          inactiveMinutes: 642,
+          lowActiveMinutes: 121,
+          mediumActiveMinutes: 14,
+          highActiveMinutes: 205,
+          lightSleepTime: 160,
+          deepSleepTime: 16,
+          timeSpentInBed: 230,
+          totalSleepTime: 205,
+          REMSleepTime: 29,
+          totalAwakeTime: 25,
+        },
+        {
+          name: "Mar-4",
+          totalCalories: 2895,
+          activeCalories: 1438,
+          totalSteps: 15727,
+          totalDistance: 14.06 * 1000,
+          inactiveMinutes: 999,
+          lowActiveMinutes: 120,
+          mediumActiveMinutes: 4,
+          highActiveMinutes: 87,
+          lightSleepTime: 253,
+          deepSleepTime: 17,
+          timeSpentInBed: 403,
+          totalSleepTime: 364,
+          REMSleepTime: 94,
+          totalAwakeTime: 39,
+        },
+        {
+          name: "Mar-5",
+          totalCalories: 2847,
+          activeCalories: 1310,
+          totalSteps: 15011,
+          totalDistance: 13.57 * 1000,
+          inactiveMinutes: 861,
+          lowActiveMinutes: 80,
+          mediumActiveMinutes: 8,
+          highActiveMinutes: 88,
+          lightSleepTime: 173,
+          deepSleepTime: 54,
+          timeSpentInBed: 298,
+          totalSleepTime: 264,
+          REMSleepTime: 0,
+          totalAwakeTime: 0,
+        },
+      ];
+      break;
+
+    case "oura":
+      result = [
+        {
+          name: "Feb-6",
+          totalCalories: 3277,
+          activeCalories: 1159,
+          totalSteps: 27800,
+          totalDistance: 22094,
+          inactiveMinutes: 532,
+          lowActiveMinutes: 142,
+          mediumActiveMinutes: 246,
+          highActiveMinutes: 0,
+          lightSleepTime: 13350 / 60,
+          deepSleepTime: 2490 / 60,
+          timeSpentInBed: 23700 / 60,
+          totalSleepTime: 21930 / 60,
+          REMSleepTime: 6090 / 60,
+          totalAwakeTime: 1770 / 60,
+          restlessSleep: 22,
+          averageHRV: 19,
+          averageRestingHR: 72.24,
+          respiratoryRate: 15.625,
+          temperatureDeviation: 0.3,
+        },
+        {
+          name: "Feb-7",
+          totalCalories: 3408,
+          activeCalories: 1249,
+          totalSteps: 15863,
+          totalDistance: 19665,
+          inactiveMinutes: 619,
+          lowActiveMinutes: 264,
+          mediumActiveMinutes: 39,
+          highActiveMinutes: 69,
+          lightSleepTime: 19680 / 60,
+          deepSleepTime: 1050 / 60,
+          timeSpentInBed: 30300 / 60,
+          totalSleepTime: 25920 / 60,
+          REMSleepTime: 5190 / 60,
+          totalAwakeTime: 4380 / 60,
+          restlessSleep: 22,
+          averageHRV: 23,
+          averageRestingHR: 64.63,
+          respiratoryRate: 16.5,
+          temperatureDeviation: 0.05,
+        },
+        {
+          name: "Feb-8",
+          totalCalories: 3312,
+          activeCalories: 1170,
+          totalSteps: 14723,
+          totalDistance: 18443,
+          inactiveMinutes: 554,
+          lowActiveMinutes: 199,
+          mediumActiveMinutes: 34,
+          highActiveMinutes: 66,
+          lightSleepTime: 13110 / 60,
+          deepSleepTime: 5430 / 60,
+          timeSpentInBed: 26820 / 60,
+          totalSleepTime: 24120 / 60,
+          REMSleepTime: 5580 / 60,
+          totalAwakeTime: 2700 / 60,
+          restlessSleep: 26,
+          averageHRV: 37,
+          averageRestingHR: 57.67,
+          respiratoryRate: 16.25,
+          temperatureDeviation: -0.17,
+        },
+        {
+          name: "Feb-9",
+          totalCalories: 3265,
+          activeCalories: 1066,
+          totalSteps: 15621,
+          totalDistance: 16693,
+          inactiveMinutes: 644,
+          lowActiveMinutes: 289,
+          mediumActiveMinutes: 34,
+          highActiveMinutes: 66,
+          lightSleepTime: 15000 / 60,
+          deepSleepTime: 2520 / 60,
+          timeSpentInBed: 24360 / 60,
+          totalSleepTime: 22380 / 60,
+          REMSleepTime: 4860 / 60,
+          totalAwakeTime: 1980 / 60,
+          restlessSleep: 29,
+          averageHRV: 34,
+          averageRestingHR: 59.58,
+          respiratoryRate: 16.5,
+          temperatureDeviation: 0.06,
+        },
+        {
+          name: "Feb-10",
+          totalCalories: 3785,
+          activeCalories: 1598,
+          totalSteps: 24851,
+          totalDistance: 26244,
+          inactiveMinutes: 560,
+          lowActiveMinutes: 237,
+          mediumActiveMinutes: 129,
+          highActiveMinutes: 70,
+          lightSleepTime: 13110 / 60,
+          deepSleepTime: 4440 / 60,
+          timeSpentInBed: 25380 / 60,
+          totalSleepTime: 23580 / 60,
+          REMSleepTime: 6030 / 60,
+          totalAwakeTime: 1800 / 60,
+          restlessSleep: 21,
+          averageHRV: 31,
+          averageRestingHR: 59.66,
+          respiratoryRate: 16.5,
+          temperatureDeviation: -0.22,
+        },
+        {
+          name: "Feb-11",
+          totalCalories: 3409,
+          activeCalories: 1239,
+          totalSteps: 15630,
+          totalDistance: 19617,
+          inactiveMinutes: 683,
+          lowActiveMinutes: 227,
+          mediumActiveMinutes: 42,
+          highActiveMinutes: 69,
+          lightSleepTime: 11880 / 60,
+          deepSleepTime: 5400 / 60,
+          timeSpentInBed: 25980 / 60,
+          totalSleepTime: 23700 / 60,
+          REMSleepTime: 6420 / 60,
+          totalAwakeTime: 2280 / 60,
+          restlessSleep: 16,
+          averageHRV: 34,
+          averageRestingHR: 59.24,
+          respiratoryRate: 16.75,
+          temperatureDeviation: 0.01,
+        },
+        {
+          name: "Feb-12",
+          totalCalories: 3263,
+          activeCalories: 1125,
+          totalSteps: 14182,
+          totalDistance: 17507,
+          inactiveMinutes: 646,
+          lowActiveMinutes: 225,
+          mediumActiveMinutes: 26,
+          highActiveMinutes: 67,
+          lightSleepTime: 10530 / 60,
+          deepSleepTime: 6600 / 60,
+          timeSpentInBed: 25980 / 60,
+          totalSleepTime: 22560 / 60,
+          REMSleepTime: 5430 / 60,
+          totalAwakeTime: 3420 / 60,
+          restlessSleep: 18,
+          averageHRV: 40,
+          averageRestingHR: 58.96,
+          respiratoryRate: 16.125,
+          temperatureDeviation: -0.09,
+        },
+        {
+          name: "Feb-13",
+          totalCalories: 3447,
+          activeCalories: 1270,
+          totalSteps: 29800,
+          totalDistance: 23673,
+          inactiveMinutes: 449,
+          lowActiveMinutes: 270,
+          mediumActiveMinutes: 247,
+          highActiveMinutes: 0,
+          lightSleepTime: 13320 / 60,
+          deepSleepTime: 2850 / 60,
+          timeSpentInBed: 25500 / 60,
+          totalSleepTime: 23730 / 60,
+          REMSleepTime: 7560 / 60,
+          totalAwakeTime: 1770 / 60,
+          restlessSleep: 19,
+          averageHRV: 22,
+          averageRestingHR: 67.64,
+          respiratoryRate: 15.5,
+          temperatureDeviation: 0.06,
+        },
+        {
+          name: "Feb-14",
+          totalCalories: 3335,
+          activeCalories: 1158,
+          totalSteps: 14760,
+          totalDistance: 17909,
+          inactiveMinutes: 593,
+          lowActiveMinutes: 242,
+          mediumActiveMinutes: 35,
+          highActiveMinutes: 66,
+          lightSleepTime: 15390 / 60,
+          deepSleepTime: 2400 / 60,
+          timeSpentInBed: 26280 / 60,
+          totalSleepTime: 23940 / 60,
+          REMSleepTime: 6150 / 60,
+          totalAwakeTime: 2340 / 60,
+          restlessSleep: 18,
+          averageHRV: 19,
+          averageRestingHR: 67.89,
+          respiratoryRate: 15.5,
+          temperatureDeviation: -0.02,
+        },
+        {
+          name: "Feb-15",
+          totalCalories: 3369,
+          activeCalories: 1176,
+          totalSteps: 15066,
+          totalDistance: 18255,
+          inactiveMinutes: 704,
+          lowActiveMinutes: 231,
+          mediumActiveMinutes: 37,
+          highActiveMinutes: 65,
+          lightSleepTime: 17160 / 60,
+          deepSleepTime: 2460 / 60,
+          timeSpentInBed: 26580 / 60,
+          totalSleepTime: 23880 / 60,
+          REMSleepTime: 4260 / 60,
+          totalAwakeTime: 2700 / 60,
+          restlessSleep: 33,
+          averageHRV: 28,
+          averageRestingHR: 62.58,
+          respiratoryRate: 16.25,
+          temperatureDeviation: -0.12,
+        },
+        {
+          name: "Feb-16",
+          totalCalories: 3335,
+          activeCalories: 1172,
+          totalSteps: 14618,
+          totalDistance: 18431,
+          inactiveMinutes: 713,
+          lowActiveMinutes: 209,
+          mediumActiveMinutes: 29,
+          highActiveMinutes: 67,
+          lightSleepTime: 12090 / 60,
+          deepSleepTime: 5670 / 60,
+          timeSpentInBed: 25020 / 60,
+          totalSleepTime: 22770 / 60,
+          REMSleepTime: 5010 / 60,
+          totalAwakeTime: 2250 / 60,
+          restlessSleep: 20,
+          averageHRV: 31,
+          averageRestingHR: 61.36,
+          respiratoryRate: 16.375,
+          temperatureDeviation: -0.08,
+        },
+        {
+          name: "Feb-17",
+          totalCalories: 3364,
+          activeCalories: 1190,
+          totalSteps: 14999,
+          totalDistance: 18634,
+          inactiveMinutes: 732,
+          lowActiveMinutes: 202,
+          mediumActiveMinutes: 31,
+          highActiveMinutes: 69,
+          lightSleepTime: 12480 / 60,
+          deepSleepTime: 5820 / 60,
+          timeSpentInBed: 25980 / 60,
+          totalSleepTime: 22860 / 60,
+          REMSleepTime: 4560 / 60,
+          totalAwakeTime: 3120 / 60,
+          restlessSleep: 19,
+          averageHRV: 44,
+          averageRestingHR: 57.39,
+          respiratoryRate: 16.375,
+          temperatureDeviation: -0.19,
+        },
+        {
+          name: "Feb-18",
+          totalCalories: 4021,
+          activeCalories: 1829,
+          totalSteps: 27651,
+          totalDistance: 31097,
+          inactiveMinutes: 606,
+          lowActiveMinutes: 158,
+          mediumActiveMinutes: 186,
+          highActiveMinutes: 72,
+          lightSleepTime: 13350 / 60,
+          deepSleepTime: 4320 / 60,
+          timeSpentInBed: 24120 / 60,
+          totalSleepTime: 22650 / 60,
+          REMSleepTime: 4980 / 60,
+          totalAwakeTime: 1470 / 60,
+          restlessSleep: 28,
+          averageHRV: 38,
+          averageRestingHR: 57.88,
+          respiratoryRate: 16.875,
+          temperatureDeviation: -0.06,
+        },
+        {
+          name: "Feb-19",
+          totalCalories: 3365,
+          activeCalories: 1207,
+          totalSteps: 15583,
+          totalDistance: 18926,
+          inactiveMinutes: 643,
+          lowActiveMinutes: 217,
+          mediumActiveMinutes: 38,
+          highActiveMinutes: 73,
+          lightSleepTime: 8940 / 60,
+          deepSleepTime: 7890 / 60,
+          timeSpentInBed: 24480 / 60,
+          totalSleepTime: 21210 / 60,
+          REMSleepTime: 4380 / 60,
+          totalAwakeTime: 3270 / 60,
+          restlessSleep: 28,
+          averageHRV: 52,
+          averageRestingHR: 54.75,
+          respiratoryRate: 16.375,
+          temperatureDeviation: -0.04,
+        },
+        {
+          name: "Feb-20",
+          totalCalories: 3341,
+          activeCalories: 1207,
+          totalSteps: 29110,
+          totalDistance: 22500,
+          inactiveMinutes: 377,
+          lowActiveMinutes: 185,
+          mediumActiveMinutes: 241,
+          highActiveMinutes: 1,
+          lightSleepTime: 10860 / 60,
+          deepSleepTime: 6030 / 60,
+          timeSpentInBed: 33000 / 60,
+          totalSleepTime: 26580 / 60,
+          REMSleepTime: 9690 / 60,
+          totalAwakeTime: 6420 / 60,
+          restlessSleep: 27,
+          averageHRV: 32,
+          averageRestingHR: 64.73,
+          respiratoryRate: 16.625,
+          temperatureDeviation: 0.37,
+        },
+        {
+          name: "Feb-21",
+          totalCalories: 3496,
+          activeCalories: 1310,
+          totalSteps: 15281,
+          totalDistance: 20330,
+          inactiveMinutes: 682,
+          lowActiveMinutes: 251,
+          mediumActiveMinutes: 30,
+          highActiveMinutes: 78,
+          lightSleepTime: 16860 / 60,
+          deepSleepTime: 2610 / 60,
+          timeSpentInBed: 29460 / 60,
+          totalSleepTime: 26670 / 60,
+          REMSleepTime: 7200 / 60,
+          totalAwakeTime: 2790 / 60,
+          restlessSleep: 25,
+          averageHRV: 24,
+          averageRestingHR: 67.18,
+          respiratoryRate: 16.5,
+          temperatureDeviation: 0.12,
+        },
+        {
+          name: "Feb-22",
+          totalCalories: 3314,
+          activeCalories: 1154,
+          totalSteps: 14337,
+          totalDistance: 18002,
+          inactiveMinutes: 736,
+          lowActiveMinutes: 190,
+          mediumActiveMinutes: 32,
+          highActiveMinutes: 68,
+          lightSleepTime: 12870 / 60,
+          deepSleepTime: 4170 / 60,
+          timeSpentInBed: 24000 / 60,
+          totalSleepTime: 20970 / 60,
+          REMSleepTime: 3930 / 60,
+          totalAwakeTime: 3030 / 60,
+          restlessSleep: 25,
+          averageHRV: 38,
+          averageRestingHR: 57.39,
+          respiratoryRate: 16.75,
+          temperatureDeviation: -0.07,
+        },
+        {
+          name: "Feb-23",
+          totalCalories: 3541,
+          activeCalories: 1372,
+          totalSteps: 16077,
+          totalDistance: 21756,
+          inactiveMinutes: 638,
+          lowActiveMinutes: 256,
+          mediumActiveMinutes: 49,
+          highActiveMinutes: 77,
+          lightSleepTime: 14160 / 60,
+          deepSleepTime: 4860 / 60,
+          timeSpentInBed: 25560 / 60,
+          totalSleepTime: 23790 / 60,
+          REMSleepTime: 4770 / 60,
+          totalAwakeTime: 1770 / 60,
+          restlessSleep: 23,
+          averageHRV: 34,
+          averageRestingHR: 60.34,
+          respiratoryRate: 17,
+          temperatureDeviation: -0.02,
+        },
+        {
+          name: "Feb-24",
+          totalCalories: 3744,
+          activeCalories: 1563,
+          totalSteps: 24466,
+          totalDistance: 25912,
+          inactiveMinutes: 596,
+          lowActiveMinutes: 192,
+          mediumActiveMinutes: 125,
+          highActiveMinutes: 68,
+          lightSleepTime: 12450 / 60,
+          deepSleepTime: 6180 / 60,
+          timeSpentInBed: 25920 / 60,
+          totalSleepTime: 23760 / 60,
+          REMSleepTime: 5130 / 60,
+          totalAwakeTime: 2160 / 60,
+          restlessSleep: 23,
+          averageHRV: 44,
+          averageRestingHR: 56.92,
+          respiratoryRate: 16.25,
+          temperatureDeviation: 0.01,
+        },
+        {
+          name: "Feb-25",
+          totalCalories: 3416,
+          activeCalories: 1253,
+          totalSteps: 14777,
+          totalDistance: 19381,
+          inactiveMinutes: 649,
+          lowActiveMinutes: 236,
+          mediumActiveMinutes: 29,
+          highActiveMinutes: 73,
+          lightSleepTime: 13170 / 60,
+          deepSleepTime: 6810 / 60,
+          timeSpentInBed: 27900 / 60,
+          totalSleepTime: 25200 / 60,
+          REMSleepTime: 5220 / 60,
+          totalAwakeTime: 2700 / 60,
+          restlessSleep: 31,
+          averageHRV: 53,
+          averageRestingHR: 56.39,
+          respiratoryRate: 15.875,
+          temperatureDeviation: -0.02,
+        },
+        {
+          name: "Feb-26",
+          totalCalories: 3356,
+          activeCalories: 1229,
+          totalSteps: 13815,
+          totalDistance: 19443,
+          inactiveMinutes: 628,
+          lowActiveMinutes: 152,
+          mediumActiveMinutes: 42,
+          highActiveMinutes: 72,
+          lightSleepTime: 12450 / 60,
+          deepSleepTime: 6810 / 60,
+          timeSpentInBed: 26340 / 60,
+          totalSleepTime: 23670 / 60,
+          REMSleepTime: 4410 / 60,
+          totalAwakeTime: 2670 / 60,
+          restlessSleep: 30,
+          averageHRV: 43,
+          averageRestingHR: 55.93,
+          respiratoryRate: 16,
+          temperatureDeviation: 0.02,
+        },
+        {
+          name: "Feb-27",
+          totalCalories: 3336,
+          activeCalories: 1175,
+          totalSteps: 29532,
+          totalDistance: 22026,
+          inactiveMinutes: 520,
+          lowActiveMinutes: 216,
+          mediumActiveMinutes: 238,
+          highActiveMinutes: 0,
+          lightSleepTime: 13050 / 60,
+          deepSleepTime: 2850 / 60,
+          timeSpentInBed: 23100 / 60,
+          totalSleepTime: 22110 / 60,
+          REMSleepTime: 6210 / 60,
+          totalAwakeTime: 990 / 60,
+          restlessSleep: 22,
+          averageHRV: 24,
+          averageRestingHR: 64.2,
+          respiratoryRate: 16.125,
+          temperatureDeviation: 0.32,
+        },
+        {
+          name: "Feb-28",
+          totalCalories: 3336,
+          activeCalories: 1187,
+          totalSteps: 14565,
+          totalDistance: 18470,
+          inactiveMinutes: 702,
+          lowActiveMinutes: 197,
+          mediumActiveMinutes: 34,
+          highActiveMinutes: 69,
+          lightSleepTime: 15810 / 60,
+          deepSleepTime: 3480 / 60,
+          timeSpentInBed: 29580 / 60,
+          totalSleepTime: 26490 / 60,
+          REMSleepTime: 7200 / 60,
+          totalAwakeTime: 3090 / 60,
+          restlessSleep: 19,
+          averageHRV: 22,
+          averageRestingHR: 64.68,
+          respiratoryRate: 16.375,
+          temperatureDeviation: 0.16,
+        },
+        {
+          name: "Mar-1",
+          totalCalories: 3391,
+          activeCalories: 1251,
+          totalSteps: 15413,
+          totalDistance: 19910,
+          inactiveMinutes: 674,
+          lowActiveMinutes: 205,
+          mediumActiveMinutes: 38,
+          highActiveMinutes: 70,
+          lightSleepTime: 15510 / 60,
+          deepSleepTime: 4410 / 60,
+          timeSpentInBed: 26580 / 60,
+          totalSleepTime: 24510 / 60,
+          REMSleepTime: 4590 / 60,
+          totalAwakeTime: 2070 / 60,
+          restlessSleep: 19,
+          averageHRV: 38,
+          averageRestingHR: 57.33,
+          respiratoryRate: 16.875,
+          temperatureDeviation: -0.25,
+        },
+        {
+          name: "Mar-2",
+          totalCalories: 3224,
+          activeCalories: 1085,
+          totalSteps: 14132,
+          totalDistance: 16976,
+          inactiveMinutes: 681,
+          lowActiveMinutes: 190,
+          mediumActiveMinutes: 32,
+          highActiveMinutes: 66,
+          lightSleepTime: 12450 / 60,
+          deepSleepTime: 5100 / 60,
+          timeSpentInBed: 27420 / 60,
+          totalSleepTime: 23760 / 60,
+          REMSleepTime: 6210 / 60,
+          totalAwakeTime: 3660 / 60,
+          restlessSleep: 25,
+          averageHRV: 47,
+          averageRestingHR: 56.53,
+          respiratoryRate: 16.875,
+          temperatureDeviation: -0.03,
+        },
+        {
+          name: "Mar-3",
+          totalCalories: 3962,
+          activeCalories: 1762,
+          totalSteps: 28454,
+          totalDistance: 29670,
+          inactiveMinutes: 609,
+          lowActiveMinutes: 201,
+          mediumActiveMinutes: 161,
+          highActiveMinutes: 72,
+          lightSleepTime: 15240 / 60,
+          deepSleepTime: 4140 / 60,
+          timeSpentInBed: 27360 / 60,
+          totalSleepTime: 24720 / 60,
+          REMSleepTime: 5340 / 60,
+          totalAwakeTime: 2640 / 60,
+          restlessSleep: 25,
+          averageHRV: 42,
+          averageRestingHR: 57.04,
+          respiratoryRate: 16.875,
+          temperatureDeviation: -0.03,
+        },
+        {
+          name: "Mar-4",
+          totalCalories: 3401,
+          activeCalories: 1248,
+          totalSteps: 15372,
+          totalDistance: 19721,
+          inactiveMinutes: 600,
+          lowActiveMinutes: 220,
+          mediumActiveMinutes: 48,
+          highActiveMinutes: 68,
+          lightSleepTime: 10560 / 60,
+          deepSleepTime: 7410 / 60,
+          timeSpentInBed: 26400 / 60,
+          totalSleepTime: 22920 / 60,
+          REMSleepTime: 4950 / 60,
+          totalAwakeTime: 3480 / 60,
+          restlessSleep: 24,
+          averageHRV: 58,
+          averageRestingHR: 53.83,
+          respiratoryRate: 15.875,
+          temperatureDeviation: -0.1,
+        },
+        {
+          name: "Mar-5",
+          totalCalories: 3469,
+          activeCalories: 1352,
+          totalSteps: 14159,
+          totalDistance: 21491,
+          inactiveMinutes: 638,
+          lowActiveMinutes: 155,
+          mediumActiveMinutes: 58,
+          highActiveMinutes: 69,
+          lightSleepTime: 11430 / 60,
+          deepSleepTime: 6450 / 60,
+          timeSpentInBed: 24720 / 60,
+          totalSleepTime: 22170 / 60,
+          REMSleepTime: 4290 / 60,
+          totalAwakeTime: 2550 / 60,
+          restlessSleep: 18,
+          averageHRV: 45,
+          averageRestingHR: 54.04,
+          respiratoryRate: 15.75,
+          temperatureDeviation: 0.03,
+        },
+      ];
+      break;
+    default:
+      break;
+  }
+  return Promise.resolve(result);
+}
