+++ packages/progress-container/src/data/helper.js
