+++ packages/progress-container/src/styles/Colors.js
@@ -1,0 +1,9 @@
+// export default Colors = {
+//     ghostWhite= '#F4F4F8',
+
+// };
+const Colors = {
+  ghostWhite: "#F4F4F8",
+};
+
+export default Colors;
