+++ packages/progress-container/src/styles/styles.css
@@ -1,0 +1,4 @@
+.progressContainer {
+  transform: rotate(90deg);
+}
+
